Move manfiest tests into module
